fix util require by moving utils to lib
