+++ lib/TvmClient.js
@@ -19,7 +19,7 @@ const tmp = require('os').tmpdir()
 const logger = require('@adobe/aio-lib-core-logging')('@adobe/aio-lib-core-tvm', { provider: 'debug' })
 const cloneDeep = require('lodash.clonedeep')
 const { codes, logAndThrow } = require('./TvmError')
-const retryRequest = require('../util/exponential-backoff')
+const retryRequest = require('./exponential-backoff')
 
 /**
  * Joins url path parts
